chore: clear pre-existing clippy and rustdoc warnings
Resolves the warnings surfaced by the new CI gates so the workspace
passes \`cargo clippy --workspace --all-targets -- -D warnings\` and
\`cargo doc --workspace --no-deps\` with \`-D warnings\`.

The bulk of the diff is mechanical clippy --fix output: collapsed
if-let chains, redundant else-break loops, redundant clones, etc.
A smaller set of edits adds targeted \`#[allow]\` attributes with
justification comments where the lint flags an intentional pattern
(trait-object \`new()\` factories, env-serialization Mutex held
across await in tests, clap-derived subcommand size disparity,
many-argument render/handler functions in the TUI/MCP layers).

Also updates two pre-existing brehon-tui snapshot baselines that
still referenced the old \`agora-tui\` / \`AGORA\` names.

+++ crates/brehon-mcp/src/tools/verification/helpers.rs
@@ -155,7 +155,7 @@ fn is_process_alive(pid: u32) -> bool {
         // all Unix platforms (macOS uses __error, Linux uses __errno_location).
         // ESRCH means the process does not exist; any other error (e.g. EPERM)
         // means the process *does* exist, so conservatively treat it as alive.
-        std::io::Error::last_os_error().raw_os_error() != Some(libc::ESRCH as i32)
+        std::io::Error::last_os_error().raw_os_error() != Some(libc::ESRCH)
     }
 }
